+++ src/containers/ticcle/create/function/checkGroupId.js
@@ -1,0 +1,7 @@
+const checkGroupId = (groupId, setTiccleGroup, initialTiccle) => {
+    (groupId === '') ? initialTiccle() : setTiccleGroup(groupId);
+}
+
+export {
+    checkGroupId
+}
