--- src/containers/ticcle/create/service/checkGroupId.js
@@ -1,14 +1,0 @@
-const checkGroupId = (groupId, setTiccleGroup, initialTiccle) => {
-    switch(groupId) {
-        case '':
-            initialTiccle();
-            break;
-        default: //screenFrom = groupId
-            setTiccleGroup(groupId);
-            break;
-    }
-}
-
-export {
-    checkGroupId
-}
