--- src/context/TiccleDetailContext.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const TiccleDetailContext = createContext();
-
-export default TiccleDetailContext;
